+++ webapp/src/main/kotlin/com/koldyr/genealogy/persistence/FamilyEventRepository.kt
@@ -1,12 +1,12 @@
 package com.koldyr.genealogy.persistence
 
-import com.koldyr.genealogy.model.FamilyEvent
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
+import com.koldyr.genealogy.model.FamilyEvent
 
 /**
  * Description of class FamilyEventRepository
  * @created: 2021-10-08
  */
 @Repository("familyEventRepository")
-interface FamilyEventRepository: JpaRepository<FamilyEvent, Int>
+interface FamilyEventRepository: JpaRepository<FamilyEvent, Long>
