+++ webapp/src/main/kotlin/com/koldyr/genealogy/persistence/PersonEventRepository.kt
@@ -1,12 +1,14 @@
 package com.koldyr.genealogy.persistence
 
-import com.koldyr.genealogy.model.PersonEvent
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
+import com.koldyr.genealogy.model.PersonEvent
 
 /**
  * Description of class FamilyEventRepository
+ *
+ * @author d.halitski@gmail.com
  * @created: 2021-10-08
  */
 @Repository("personEventRepository")
-interface PersonEventRepository: JpaRepository<PersonEvent, Int>
+interface PersonEventRepository: JpaRepository<PersonEvent, Long>
